Fix import project from sending request before runtime is active
